Update check out 1.5 seconds after an email is changed.
Similar to how WC core updates the checkout after an address field is
changed. This gets around the loophole where the purchaser may not
unfocus the email field before processing the checkout.

Also, remove storing recipients in the session. This is no longer
necessary now that we're setting them on the cart item directly.