Tidy up the JS ensuring the checkout isn't reloaded more than necessary
If you changed the recipient and let the timer reset the checkout this
would also cause the change event to trigger when the old field was
removed. To get around that this commit introduces the 'needs_update'
class which is applied to elements which need to trigger a checkout
update.