Don't reformat email subjects and headings post WC 3.1
WC 3.1 removed the dynamic email headings and subjects for emails which
contained download links. In order for the correct headings to be sent
to recipients and purchasers, Gifting would determine whether the email
recipient had downloads and if they didn't, would use the
`$email->heading` or `$email->subject` values to get the default values.
These were removed though and caused blank email subjects to be
returned.

This commit completely removes that process for sites running 3.1 + as
it's not necessary.